backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
python-multipart>=0.0.9
anthropic>=0.40.0
python-dotenv>=1.0.1
